moved logic for dynamic color change to the topLevel app component
