+++ console/src/views/ThemeSettings.js
@@ -8,12 +8,11 @@ import kind from '@enact/core/kind';
 import {Row, Column, Cell} from '@enact/ui/Layout';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
-import {useCallback, useContext, useEffect, useMemo, useState} from 'react';
+import {useCallback, useContext, useState} from 'react';
 
 import {getPanelIndexOf} from '../App';
 import AppContextConnect from '../App/AppContextConnect';
 import {AppContext} from '../App/AppContextProvider';
-import {generateTimestamps, getColorsDayMode, getColorsNightMode, getIndex} from '../utils';
 
 import componentCss from './Settings.module.less';
 
@@ -29,7 +28,6 @@ const skinCollection = {
 };
 const skinList = Object.keys(skinCollection);
 const skinNames = skinList.map((skin) => skinCollection[skin]);  // Build the names list based on the skin list array, so the indexes always match up, in case the object keys don't return in the same order.
-let fakeIndex = 0;
 const skinVariantsCollection = {
 	'': 'Day',
 	night: 'Night'
@@ -44,7 +42,6 @@ const swatchPalette = [
 	'#f08c21', '#f42c04', '#d636d9', '#aab3cb',
 	'#ff2d55', '#9e00d8'
 ];
-const timestamps = generateTimestamps(5);
 
 const FormRow = kind({
 	name: 'FormRow',
@@ -117,7 +114,7 @@ const ColorCheckboxItem = kind({
 
 const ThemeSettingsBase = (props) => {
 	const [realTime, setRealTime] = useState(false);
-	const {prevIndex, ... rest} = props;
+	const {onToggleDynamicColor, prevIndex,  ... rest} = props;
 	const context = useContext(AppContext);
 
 	delete rest.onSendSkinSettings;
@@ -135,101 +132,6 @@ const ThemeSettingsBase = (props) => {
 	const onChange = useCallback(() => {
 		props.onSendSkinSettings(props);
 	}, [props]);
-
-	const accentColors = {};
-	const highlightColors = {};
-
-	const accentColorsArray = useMemo(() => {
-		const dayColorArray = getColorsDayMode(context.userSettings.colorAccentManual, 72);
-		const nightColorArray = getColorsNightMode(context.userSettings.colorAccentManual, 72);
-		const array = [...nightColorArray.reverse(), ...dayColorArray, ...dayColorArray.reverse(), ...nightColorArray.reverse()];
-		const offset = array.splice(0, 12);
-
-		return [...array, ...offset];
-	}, [context.userSettings.colorAccentManual]);
-
-	const highlightColorsArray = useMemo(() => {
-		const dayColorArray = getColorsDayMode(context.userSettings.colorHighlightManual, 72);
-		const nightColorArray = getColorsNightMode(context.userSettings.colorHighlightManual, 72);
-		const array = [...dayColorArray.reverse(), ...nightColorArray, ...nightColorArray.reverse(), ...dayColorArray.reverse()];
-		const offset = array.splice(0, 12);
-
-		return [...array, ...offset];
-	}, [context.userSettings.colorHighlightManual]);
-
-	timestamps.forEach((element, index) => {
-		accentColors[element] = accentColorsArray[index];
-	});
-
-	timestamps.forEach((element, index) => {
-		highlightColors[element] = highlightColorsArray[index];
-	});
-
-	useEffect (() => {
-		let changeColor;
-
-		if (dynamicColorActive) {
-			// Remove this when fake timers will no longer be used
-			context.updateAppState((state) => {
-				state.userSettings.colorAccent = context.userSettings.colorAccentManual;
-				state.userSettings.colorHighlight = context.userSettings.colorHighlightManual;
-				state.userSettings.skinVariants = context.userSettings.skinVariantsManual;
-			});
-
-			context.updateAppState((state) => {
-				state.userSettings.colorAccentManual = state.userSettings.colorAccent;
-				state.userSettings.colorHighlightManual = state.userSettings.colorHighlight;
-				state.userSettings.skinVariantsManual = state.userSettings.skinVariants;
-			});
-
-			changeColor = setInterval(() => {
-				if (realTime) {
-					const index = getIndex();
-					let skinVariant;
-					if (index >= '06:00' && index < '18:00') {
-						skinVariant = '';
-					} else {
-						skinVariant = 'night';
-					}
-
-					context.updateAppState((state) => {
-						state.userSettings.colorAccent = `${accentColors[index]}`;
-						state.userSettings.colorHighlight = `${highlightColors[index]}`;
-						state.userSettings.skinVariants = skinVariant;
-					});
-				} else {
-					let skinVariant;
-					if (60 <= fakeIndex && fakeIndex <= 203) {
-						skinVariant = '';
-					} else {
-						skinVariant = 'night';
-					}
-
-					context.updateAppState((state) => {
-						state.userSettings.colorAccent = `${accentColorsArray[fakeIndex]}`;
-						state.userSettings.colorHighlight = `${highlightColorsArray[fakeIndex]}`;
-						state.userSettings.skinVariants = skinVariant;
-					});
-					if (fakeIndex < 287) {
-						fakeIndex++;
-					} else {
-						fakeIndex = 0;
-					}
-				}
-			}, realTime ? 30 * 1000 : 100);
-		} else {
-			context.updateAppState((state) => {
-				state.userSettings.colorAccent = context.userSettings.colorAccentManual;
-				state.userSettings.colorHighlight = context.userSettings.colorHighlightManual;
-				state.userSettings.skinVariants = context.userSettings.skinVariantsManual;
-			});
-		}
-
-		return () => {
-			clearInterval(changeColor);
-			fakeIndex = 0;
-		};
-	}, [context.userSettings.dynamicColor, realTime]); // eslint-disable-line react-hooks/exhaustive-deps
 
 	return (
 		<Panel {...rest} className={componentCss.settingsView}>
@@ -279,7 +181,7 @@ const ThemeSettingsBase = (props) => {
 							</FontSizeSetting>
 						</Cell>*/}
 						<Cell shrink className={componentCss.spacedItem}>
-							<DynamicColorSetting>
+							<DynamicColorSetting onToggle={onToggleDynamicColor}>
 								Dynamic color change
 							</DynamicColorSetting>
 							<ColorCheckboxItem onToggle={handleRealState} value={realTime}>
@@ -299,6 +201,7 @@ ThemeSettingsBase.propTypes = {
 	label: PropTypes.string,
 	onSelect: PropTypes.func,
 	onSendSkinSettings: PropTypes.func,
+	onToggleDynamicColor: PropTypes.func,
 	prevIndex: PropTypes.number,
 	skin: PropTypes.string
 };
@@ -345,13 +248,8 @@ const SkinVariantsSetting = AppContextConnect(({userSettings, updateAppState}) =
 	}
 }))(SliderButtonItem);
 
-const DynamicColorSetting = AppContextConnect(({userSettings, updateAppState}) => ({
-	selected: userSettings.dynamicColor,
-	onToggle: ({selected}) => {
-		updateAppState((state) => {
-			state.userSettings.dynamicColor = selected;
-		});
-	}
+const DynamicColorSetting = AppContextConnect(({userSettings}) => ({
+	selected: userSettings.dynamicColor
 }))(ColorCheckboxItem);
 
 const ThemeSettingsDecorator = compose(
